x/raytracing: modify code to support batch raytracing
individual rays trace about 10 times slower, which is a shame.
But batch raytracing does 1M rays in 796ms for two surfaces
(~400ms/surf => 2.5M ray-surfaces /sec).

Prior impl did ~20k ray-surfaces/sec, this is a 125x speedup

There may be gains on the table with BLAS numpy functions replacing
the sum(a*b, axis=1) calls and a few other places, but this is quite
good already

A few codependent tweaks are needed in the surfaces type
(specifically, the plane factory)

An immature set of fixes exist in an ipynb but need to be modified for
dtype stability and good use with scalars

the phist and shist need a squeeze when there is only one ray, a future
commit will add that