basic jones calculus + tests
wrote some functions to construct jones matrices with the option to broadcast them in arbitrary dimensions. Also added some tests for the functions.

The option to convert them to Mueller matrices is included, and the Pauli spin matrices are added to play with polarization aberrations later.

making a PR to experimental as a place to discuss the "right way" to structure such a code.